+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### General Changes
 * Updated Alpine Linux version for static builds to latest v3.22.
 * Updated mimalloc lib for static builds to latest v2.2.4.
+* Let service instances respond to `--quit` command so that caller can see which services were running.
 
 ### Fixes
 * For some bucket-related phases, the performance report on the console had "Objects/s" instead of "Buckets/s".